create model CurrentUser
